- AUT-340 	Ensure we deal with multiple data attributes - improvements to handling exceptions set in jni code

+++ bindings/java/src/org/sleuthkit/datamodel/Directory.java
@@ -31,7 +31,7 @@ public class Directory extends FsContent {
 
 	//constructor used for getdir from tskDb
 	protected Directory(SleuthkitCase db, long obj_id, long fs_obj_id, long meta_addr,
-			long attr_type, long attr_id, String name, long dir_type,
+			int attr_type, int attr_id, String name, long dir_type,
 			long meta_type, long dir_flags, long meta_flags, long size,
 			long ctime, long crtime, long atime, long mtime, long mode,
 			long uid, long gid, long known, String parent_path, String md5Hash) {
